signup-success as own route. Removed base Signup Vue View.
