--- src/views/Signup.vue
@@ -1,34 +1,0 @@
-<template>
-        <div class="container">
-        <small-page>
-            <router-view/>
-        </small-page>
-        <a class="global-close" :class="{hidden: $route.name === `signup-success`}" @click="close">Back to {{request.appName}}</a>
-    </div>
-</template>
-
-<script lang="ts">
-import {Component, Emit, Prop, Watch, Vue} from 'vue-property-decorator';
-import {SmallPage} from '@nimiq/vue-components';
-import {RequestType, ParsedSignupRequest} from '../lib/RequestTypes';
-import {AddressInfo} from '../lib/AddressInfo';
-import {KeyInfo, KeyStorageType} from '../lib/KeyInfo';
-import {State, Mutation, Getter} from 'vuex-class';
-import RpcApi from '../lib/RpcApi';
-import {CreateRequest as KCreateRequest, CreateResult as KCreateResult} from '@nimiq/keyguard-client';
-import {ResponseStatus, State as RpcState} from '@nimiq/rpc';
-import {Static} from '../lib/StaticStore';
-
-@Component({components: {SmallPage}})
-export default class extends Vue {
-    @Static private rpcState!: RpcState;
-    @Static private request!: ParsedSignupRequest;
-    @State private keyguardResult!: KCreateResult | Error | null;
-    @State private activeAccountPath!: string;
-
-    @Emit()
-    private close() {
-        this.rpcState.reply(ResponseStatus.ERROR, new Error('CANCEL'));
-    }
-}
-</script>
