+++ src/views/SignupSuccess.vue
@@ -1,38 +1,43 @@
 <template>
-    <div>
-        <PageHeader :progressIndicator="true" :numberSteps="6" :step="6">Your wallet is ready</PageHeader>
-        <div class="page-body">
-            <div class="success-box">
-                <h2>Awesome!</h2>
-                <p>Your Keyguard Wallet is set up. It already contains your newly created account.</p>
-                <p>You can add more accounts to it later.</p>
+    <div class="container">
+        <small-page>
+            <div>
+                <PageHeader :progressIndicator="true" :numberSteps="6" :step="6">Your wallet is ready</PageHeader>
+                <div class="page-body">
+                    <div class="success-box">
+                        <h2>Awesome!</h2>
+                        <p>Your Keyguard Wallet is set up. It already contains your newly created account.</p>
+                        <p>You can add more accounts to it later.</p>
+                    </div>
+
+                    <div class="login-label">
+                        <div class="login-icon" :class="walletIconClass"></div>
+                        <LabelInput :value="walletLabel" @changed="onWalletLabelChange"/>
+                    </div>
+
+                    <Account :address="createdAddress" :label="accountLabel" :editable="true" @changed="onAccountLabelChange"/>
+
+                    <button class="submit" @click="done()">Open your wallet</button>
+                </div>
             </div>
-
-            <div class="login-label">
-                <div class="login-icon" :class="walletIconClass"></div>
-                <LabelInput :value="walletLabel" @changed="onWalletLabelChange"/>
-            </div>
-
-            <Account :address="createdAddress" :label="accountLabel" :editable="true" @changed="onAccountLabelChange"/>
-
-            <button class="submit" @click="done()">Open your wallet</button>
-        </div>
+        </small-page>
     </div>
+
 </template>
 
 <script lang="ts">
-import {Component, Vue} from 'vue-property-decorator';
-import {PageHeader, Account, LabelInput} from '@nimiq/vue-components';
-import {AddressInfo} from '../lib/AddressInfo';
-import {KeyInfo, KeyStorageType} from '../lib/KeyInfo';
-import {State, Getter} from 'vuex-class';
-import {KeyStore} from '../lib/KeyStore';
-import {CreateResult} from '@nimiq/keyguard-client';
-import {ResponseStatus, State as RpcState} from '@nimiq/rpc';
+import { Component, Vue } from 'vue-property-decorator';
+import { PageHeader, Account, LabelInput, SmallPage } from '@nimiq/vue-components';
+import { AddressInfo } from '../lib/AddressInfo';
+import { KeyInfo, KeyStorageType } from '../lib/KeyInfo';
+import { State, Getter } from 'vuex-class';
+import { KeyStore } from '../lib/KeyStore';
+import { CreateResult } from '@nimiq/keyguard-client';
+import { ResponseStatus, State as RpcState } from '@nimiq/rpc';
 import { SignupResult } from '@/lib/RequestTypes';
-import {Static} from '../lib/StaticStore';
+import { Static } from '../lib/StaticStore';
 
-@Component({components: {PageHeader, Account, LabelInput}})
+@Component({components: {PageHeader, Account, LabelInput, SmallPage}})
 export default class extends Vue {
     @Static private rpcState!: RpcState;
     @State private keyguardResult!: CreateResult;
